Use webpack
To run the code in the browser, imports are required to be to ".js" files
including the extension. However, Mocha can't seem to understand ES6 modules
and adding the .js extension to the imports breaks it.

Using webpack allows me to have code the Mocha understands, and then bundle it
into a format with imports that browser will understand. Ugh. This is so
convoluted.