Remove svgo convertStyleToAttrs and use !important css instead
Now that I'm using embedded CSS in the svg convertStyleToAttrs changes
how it renders since an styles that differ from the base class are
converted to attrs which are overridden by the class.

This fixes some issues caused by adding the datapath-label SVG CSS

!important actually allows CSS to override inline styles, so we don't
actually need convertStyleToAttrs. I should have done that to begin
with.

This does cause a problem with  the markers though. Markers no longer
being an attribute means I can't css select them. So I'll have to fix
that next.